+++ packages/oauth-client/tests/unit/jest.config.ts
@@ -1,0 +1,12 @@
+import { jestUnitConfig } from '@ttoss/config';
+
+export default jestUnitConfig({
+  coverageThreshold: {
+    global: {
+      statements: 99,
+      branches: 99,
+      functions: 99,
+      lines: 99,
+    },
+  },
+});