+++ challenge/packages/app-vite-with-graphql-api/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.31](https://github.com/ttoss/ttoss/compare/@ttoss-challenge/app-vite-with-graphql-api@0.4.30...@ttoss-challenge/app-vite-with-graphql-api@0.4.31) (2026-06-12)
+
+**Note:** Version bump only for package @ttoss-challenge/app-vite-with-graphql-api
+
 ## [0.4.30](https://github.com/ttoss/ttoss/compare/@ttoss-challenge/app-vite-with-graphql-api@0.4.29...@ttoss-challenge/app-vite-with-graphql-api@0.4.30) (2026-06-10)
 
 **Note:** Version bump only for package @ttoss-challenge/app-vite-with-graphql-api