+++ docs/storybook/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.25.5](https://github.com/ttoss/ttoss/compare/@docs/storybook@3.25.4...@docs/storybook@3.25.5) (2026-06-12)
+
+**Note:** Version bump only for package @docs/storybook
+
 ## [3.25.4](https://github.com/ttoss/ttoss/compare/@docs/storybook@3.25.3...@docs/storybook@3.25.4) (2026-06-10)
 
 **Note:** Version bump only for package @docs/storybook