+++ packages/http-server-auth/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.3](https://github.com/ttoss/ttoss/compare/@ttoss/http-server-auth@0.3.2...@ttoss/http-server-auth@0.3.3) (2026-06-14)
+
+**Note:** Version bump only for package @ttoss/http-server-auth
+
+## [0.3.2](https://github.com/ttoss/ttoss/compare/@ttoss/http-server-auth@0.3.1...@ttoss/http-server-auth@0.3.2) (2026-06-14)
+
+**Note:** Version bump only for package @ttoss/http-server-auth
+
+## [0.3.1](https://github.com/ttoss/ttoss/compare/@ttoss/http-server-auth@0.3.0...@ttoss/http-server-auth@0.3.1) (2026-06-14)
+
+**Note:** Version bump only for package @ttoss/http-server-auth
+
+# [0.3.0](https://github.com/ttoss/ttoss/compare/@ttoss/http-server-auth@0.2.1...@ttoss/http-server-auth@0.3.0) (2026-06-14)
+
+### Features
+
+- **http-server-auth:** pass ctx to auth callbacks + MCP OAuth guideline ([#1061](https://github.com/ttoss/ttoss/issues/1061)) ([c627241](https://github.com/ttoss/ttoss/commit/c627241748d27d93849cb81569d08e0773a466f3)), closes [#1058](https://github.com/ttoss/ttoss/issues/1058)
+
 ## [0.2.1](https://github.com/ttoss/ttoss/compare/@ttoss/http-server-auth@0.2.0...@ttoss/http-server-auth@0.2.1) (2026-06-11)
 
 **Note:** Version bump only for package @ttoss/http-server-auth